+++ CHANGES.md
@@ -18,6 +18,7 @@ Features
   (CONNECTED) /> ls / watch=true
   ```
 - new command `time`: allows time the execution of the given commands
+- the `create` command now supports async mode (async=true)
 
 1.1.0 (2015-06-17)
 --------------------